Began separation of database logic.
All logic involving the database has been moved from server.js into
dao.js. server.js needs to be changed to accept return values from the
DAO, and the DAO needs additional code to properly return values from
its functions.
