+++ .gitignore
@@ -2,4 +2,3 @@
 node_modules/
 .DS_Store
 query_test.js
-dao.js
